+++ cli/src/test/java/org/kosit/validator/impl/tasks/TestScenarioBuilder.java
@@ -7,10 +7,10 @@ import org.kosit.validator.impl.ResolvingMode;
 import org.kosit.validator.impl.Scenario;
 import org.kosit.validator.impl.xml.ProcessorProvider;
 import org.kosit.validator.impl.xml.StrictRelativeResolvingStrategy;
-import org.kosit.validator.model.scenarios.CreateReportType;
-import org.kosit.validator.model.scenarios.ResourceType;
-import org.kosit.validator.model.scenarios.ScenarioType;
-import org.kosit.validator.model.scenarios.ValidateWithXmlSchema;
+import org.kosit.validator.scenario.model.CreateReportType;
+import org.kosit.validator.scenario.model.ResourceType;
+import org.kosit.validator.scenario.model.ScenarioType;
+import org.kosit.validator.scenario.model.ValidateWithXmlSchema;
 
 import javax.xml.validation.Schema;
 import java.io.IOException;