+++ core/src/test/java/org/kosit/validator/config/SchemaBuilderTest.java
@@ -12,8 +12,8 @@ import org.junit.jupiter.api.Test;
 import org.kosit.validator.impl.ContentRepository;
 import org.kosit.validator.impl.TestHelper.Simple;
 import org.kosit.validator.impl.model.Result;
-import org.kosit.validator.model.scenarios.ResourceType;
-import org.kosit.validator.model.scenarios.ValidateWithXmlSchema;
+import org.kosit.validator.scenario.model.ResourceType;
+import org.kosit.validator.scenario.model.ValidateWithXmlSchema;
 
 /**
  * Tests {@link SchemaBuilder}.