+++ core/src/test/java/org/kosit/validator/impl/tasks/SchematronValidationTaskTest.java
@@ -21,8 +21,8 @@ import org.kosit.validator.impl.TestHelper.Simple;
 import org.kosit.validator.impl.model.ProcessStepResult;
 import org.kosit.validator.impl.model.Result;
 import org.kosit.validator.model.ValidationResultsSchematron;
-import org.kosit.validator.model.scenarios.ResourceType;
-import org.kosit.validator.model.scenarios.ValidateWithSchematron;
+import org.kosit.validator.scenario.model.ResourceType;
+import org.kosit.validator.scenario.model.ValidateWithSchematron;
 
 import net.sf.saxon.s9api.SaxonApiException;
 import net.sf.saxon.s9api.XsltExecutable;